requirements.txt
matplotlib==3.6.2
numpy==1.24.1
Pillow==9.3.0
polars==0.15.10
requests==2.28.2